requirements.txt
wandb
torch
torchaudio
matplotlib
numpy
librosa
python-dotenv
jupyterlab
gradio